RFC: goodbye inheritance, hello composition
